requirements.txt
fastapi
uvicorn
httpx
black
jinja2
aiofiles
python-dotenv
requests